Increase GC interval to 3 minutes
There is some need to be careful here though. The garbage collection will actually run at GCinterval
+ GCdelay, which leads to 4 minutes.

(When the server is initially started, the lastGCrun can be set to as old a timestamp as 4 minutes
away. This is not a problem in producion; it is only an issue in development, where we want to get
the GC to run immediately and check its impact. So, the first time the GC runs will be something
like 8-10 minutes after the server has started. With my other current configurations, there are only
10 overTime slots, so this was pretty close to the case where GC did not run and all overTime slots
being used up.)